pyproject.toml
[tool.poetry]
name = "python-tutorial"
version = "0.1.0"
description = ""
authors = ["alexzheng111 <alexzhengalt@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^8.3.4"
fastapi = {extras = ["standard"], version = "^0.115.6"}
pre-commit = "^4.0.1"
flake8 = "^7.1.1"
black = "^24.10.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
